fix: process_id를 BIGINT로 변경하여 JavaScript 타임스탬프 ID 지원
- 문제: JavaScript 타임스탬프 기반 ID(1756946076323)가 PostgreSQL int32 범위 초과
- 해결: 모든 process_id 관련 컬럼을 Integer에서 BigInteger로 변경
- 수정된 엔티티:
  - Process.id: Integer  BigInteger
  - ProductProcess.process_id: Integer  BigInteger
  - MatDir.process_id: Integer  BigInteger
  - FuelDir.process_id: Integer  BigInteger
  - Edge.source_id, target_id: Integer  BigInteger
  - ProcessAttrdirEmission.process_id: Integer  BigInteger
- 마이그레이션 스크립트 추가: fix_process_id_bigint.sql
- 이제 JavaScript에서 생성된 큰 숫자 ID도 정상 처리 가능

+++ service/cbam_service/app/domain/calculation/calculation_entity.py
@@ -21,7 +21,7 @@ class ProcessAttrdirEmission(Base):
     __tablename__ = "process_attrdir_emission"
     
     id = Column(Integer, primary_key=True, index=True)
-    process_id = Column(Integer, ForeignKey("process.id", ondelete="CASCADE"), nullable=False, index=True)
+    process_id = Column(BigInteger, ForeignKey("process.id", ondelete="CASCADE"), nullable=False, index=True)
     total_matdir_emission = Column(Numeric(15, 6), nullable=False, default=0, comment="총 원료직접배출량")
     total_fueldir_emission = Column(Numeric(15, 6), nullable=False, default=0, comment="총 연료직접배출량")
     attrdir_em = Column(Numeric(15, 6), nullable=False, default=0, comment="직접귀속배출량 (원료+연료)")
